upadate front

+++ paq-backend/src/main/java/com/polytech/paqbackend/token/TokenRepository.java
@@ -5,15 +5,18 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 import java.util.Optional;
-public interface TokenRepository  extends JpaRepository<Token, Integer> {
+public interface TokenRepository  extends JpaRepository<Token, Long> {
     @Query("""
-            select t from Token t inner join User u
-            on t.user.id = u.id
-            where u.id = :id and (t.expired = false or t.revoked = false)
-            """)
+    select t from Token t inner join User u
+    on t.user.id = u.id
+    where u.id = :id and (t.expired = false or t.revoked = false)
+    """)
     List<Token> findAllValidTokenByUser(Long id);
 
     Optional<Token> findByToken(String token);
+
+
+
 
 
 }
